Adding a RawQueryoperation
In moving from Scanamo 0.9.2 to 1.0.0 we have found that there is no longer support for a Java based QueryRequest.  I *believe* the intent behind that is to force people to use the ScanamoResult.  However, for us this causes limitations as we have queries that are not supported with the grammar provided (we believe).  So this PR is in the spirit of supporting a raw Java QueryRequest and interpreting it.

More work needs to be done to add proper tests but this is an attempt at starting the conversation and going from there.